Modifications - data-table is now more modular - Task Queries and Mutations present in app/(tasks) - None present inside data-table - Column Defs are now present in tasks-table-column-def - Actions on Selected Rows are in tasks-table-selected-action-controls - useDataTable Hook present in `@/hooks/use-data-table.ts - useDataTable abstracts the logic from <DataTable> component - More reusable code
