+++ src/app/(tasks)/_actions/validations.ts
@@ -2,16 +2,16 @@ import { tasks } from "@/db/schema"
 import * as z from "zod"
 
 export const updateTaskLabelSchema = z.object({
-  id: z.number(),
+  uid: z.number(),
   label: z.enum(tasks.label.enumValues),
 })
 
 export const updateTaskStatusSchema = z.object({
-  id: z.number(),
+  uid: z.number(),
   status: z.enum(tasks.status.enumValues),
 })
 
 export const updateTaskPrioritySchema = z.object({
-  id: z.number(),
+  uid: z.number(),
   priority: z.enum(tasks.priority.enumValues),
 })
