+++ src/components/data-table/data-table.tsx
@@ -1,26 +1,15 @@
-import * as React from "react"
-import { usePathname, useRouter, useSearchParams } from "next/navigation"
 import type {
   DataTableFilterableColumn,
   DataTableSearchableColumn,
-} from "@/types"
+} from "@/types";
+import * as React from "react";
+
 import {
   flexRender,
-  getCoreRowModel,
-  getFacetedRowModel,
-  getFacetedUniqueValues,
-  getFilteredRowModel,
-  getPaginationRowModel,
-  getSortedRowModel,
-  useReactTable,
+  type Table as TanstackTable,
   type ColumnDef,
-  type ColumnFiltersState,
-  type PaginationState,
-  type SortingState,
-  type VisibilityState,
-} from "@tanstack/react-table"
+} from "@tanstack/react-table";
 
-import { useDebounce } from "@/hooks/use-debounce"
 import {
   Table,
   TableBody,
@@ -28,251 +17,55 @@ import {
   TableHead,
   TableHeader,
   TableRow,
-} from "@/components/ui/table"
+} from "@/components/ui/table";
 
-import { DataTableAdvancedToolbar } from "./advanced/data-table-advanced-toolbar"
-import { DataTableFloatingBar } from "./data-table-floating-bar"
-import { DataTablePagination } from "./data-table-pagination"
-import { DataTableToolbar } from "./data-table-toolbar"
+import { DataTableAdvancedToolbar } from "./advanced/data-table-advanced-toolbar";
+import { DataTablePagination } from "./data-table-pagination";
+import { DataTableFloatingBar } from "./data-table-floating-bar";
+import { DataTableToolbar } from "./data-table-toolbar";
 
-interface DataTableProps<TData, TValue> {
+interface DataTableProps<TData extends { uid: number }, TValue> {
+  dataTable: TanstackTable<TData>
   columns: ColumnDef<TData, TValue>[]
-  data: TData[]
-  pageCount: number
   filterableColumns?: DataTableFilterableColumn<TData>[]
   searchableColumns?: DataTableSearchableColumn<TData>[]
   advancedFilter?: boolean
   floatingBar?: boolean
-  deleteRowsAction?: React.MouseEventHandler<HTMLButtonElement>
+  selectedActionControls: React.ReactNode;
+  deleteRowsAction: React.MouseEventHandler<HTMLButtonElement>
 }
 
-export function DataTable<TData, TValue>({
+export function DataTable<TData extends { uid: number }, TValue>({
+  dataTable,
   columns,
-  data,
-  pageCount,
   filterableColumns = [],
   searchableColumns = [],
   advancedFilter = false,
   floatingBar = false,
+  selectedActionControls,
   deleteRowsAction,
 }: DataTableProps<TData, TValue>) {
-  const router = useRouter()
-  const pathname = usePathname()
-  const searchParams = useSearchParams()
-
-  // Search params
-  const page = searchParams?.get("page") ?? "1"
-  const pageAsNumber = Number(page)
-  const fallbackPage =
-    isNaN(pageAsNumber) || pageAsNumber < 1 ? 1 : pageAsNumber
-  const per_page = searchParams?.get("per_page") ?? "10"
-  const perPageAsNumber = Number(per_page)
-  const fallbackPerPage = isNaN(perPageAsNumber) ? 10 : perPageAsNumber
-  const sort = searchParams?.get("sort")
-  const [column, order] = sort?.split(".") ?? []
-
-  // Create query string
-  const createQueryString = React.useCallback(
-    (params: Record<string, string | number | null>) => {
-      const newSearchParams = new URLSearchParams(searchParams?.toString())
-
-      for (const [key, value] of Object.entries(params)) {
-        if (value === null) {
-          newSearchParams.delete(key)
-        } else {
-          newSearchParams.set(key, String(value))
-        }
-      }
-
-      return newSearchParams.toString()
-    },
-    [searchParams]
-  )
-
-  // Table states
-  const [rowSelection, setRowSelection] = React.useState({})
-  const [columnVisibility, setColumnVisibility] =
-    React.useState<VisibilityState>({})
-  const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>(
-    []
-  )
-
-  // Handle server-side pagination
-  const [{ pageIndex, pageSize }, setPagination] =
-    React.useState<PaginationState>({
-      pageIndex: fallbackPage - 1,
-      pageSize: fallbackPerPage,
-    })
-
-  const pagination = React.useMemo(
-    () => ({
-      pageIndex,
-      pageSize,
-    }),
-    [pageIndex, pageSize]
-  )
-
-  React.useEffect(() => {
-    setPagination({
-      pageIndex: fallbackPage - 1,
-      pageSize: fallbackPerPage,
-    })
-  }, [fallbackPage, fallbackPerPage])
-
-  React.useEffect(() => {
-    router.push(
-      `${pathname}?${createQueryString({
-        page: pageIndex + 1,
-        per_page: pageSize,
-      })}`,
-      {
-        scroll: false,
-      }
-    )
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [pageIndex, pageSize])
-
-  // Handle server-side sorting
-  const [sorting, setSorting] = React.useState<SortingState>([
-    {
-      id: column ?? "",
-      desc: order === "desc",
-    },
-  ])
-
-  React.useEffect(() => {
-    router.push(
-      `${pathname}?${createQueryString({
-        page,
-        sort: sorting[0]?.id
-          ? `${sorting[0]?.id}.${sorting[0]?.desc ? "desc" : "asc"}`
-          : null,
-      })}`
-    )
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [sorting])
-
-  // Handle server-side filtering
-  const debouncedSearchableColumnFilters = JSON.parse(
-    useDebounce(
-      JSON.stringify(
-        columnFilters.filter((filter) => {
-          return searchableColumns.find((column) => column.id === filter.id)
-        })
-      ),
-      500
-    )
-  ) as ColumnFiltersState
-
-  const filterableColumnFilters = columnFilters.filter((filter) => {
-    return filterableColumns.find((column) => column.id === filter.id)
-  })
-
-  React.useEffect(() => {
-    for (const column of debouncedSearchableColumnFilters) {
-      if (typeof column.value === "string") {
-        router.push(
-          `${pathname}?${createQueryString({
-            page: 1,
-            [column.id]: typeof column.value === "string" ? column.value : null,
-          })}`
-        )
-      }
-    }
-
-    for (const key of searchParams.keys()) {
-      if (
-        searchableColumns.find((column) => column.id === key) &&
-        !debouncedSearchableColumnFilters.find((column) => column.id === key)
-      ) {
-        router.push(
-          `${pathname}?${createQueryString({
-            page: 1,
-            [key]: null,
-          })}`
-        )
-      }
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [JSON.stringify(debouncedSearchableColumnFilters)])
-
-  React.useEffect(() => {
-    for (const column of filterableColumnFilters) {
-      if (typeof column.value === "object" && Array.isArray(column.value)) {
-        router.push(
-          `${pathname}?${createQueryString({
-            page: 1,
-            [column.id]: column.value.join("."),
-          })}`
-        )
-      }
-    }
-
-    for (const key of searchParams.keys()) {
-      if (
-        filterableColumns.find((column) => column.id === key) &&
-        !filterableColumnFilters.find((column) => column.id === key)
-      ) {
-        router.push(
-          `${pathname}?${createQueryString({
-            page: 1,
-            [key]: null,
-          })}`
-        )
-      }
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [JSON.stringify(filterableColumnFilters)])
-
-  const table = useReactTable({
-    data,
-    columns,
-    pageCount: pageCount ?? -1,
-    state: {
-      pagination,
-      sorting,
-      columnVisibility,
-      rowSelection,
-      columnFilters,
-    },
-    enableRowSelection: true,
-    onRowSelectionChange: setRowSelection,
-    onPaginationChange: setPagination,
-    onSortingChange: setSorting,
-    onColumnFiltersChange: setColumnFilters,
-    onColumnVisibilityChange: setColumnVisibility,
-    getCoreRowModel: getCoreRowModel(),
-    getFilteredRowModel: getFilteredRowModel(),
-    getPaginationRowModel: getPaginationRowModel(),
-    getSortedRowModel: getSortedRowModel(),
-    getFacetedRowModel: getFacetedRowModel(),
-    getFacetedUniqueValues: getFacetedUniqueValues(),
-    manualPagination: true,
-    manualSorting: true,
-    manualFiltering: true,
-  })
 
   return (
     <div className="w-full space-y-2.5 overflow-auto">
       {advancedFilter ? (
         <DataTableAdvancedToolbar
-          table={table}
+          dataTable={dataTable}
           filterableColumns={filterableColumns}
           searchableColumns={searchableColumns}
         />
       ) : (
         <DataTableToolbar
-          table={table}
+          table={dataTable}
           filterableColumns={filterableColumns}
           searchableColumns={searchableColumns}
+          deleteRowsAction={floatingBar ? undefined :  deleteRowsAction}
         />
       )}
       <div className="rounded-md border">
         <Table>
           <TableHeader>
-            {table.getHeaderGroups().map((headerGroup) => (
+            {dataTable.getHeaderGroups().map((headerGroup) => (
               <TableRow key={headerGroup.id}>
                 {headerGroup.headers.map((header) => {
                   return (
@@ -280,9 +73,9 @@ export function DataTable<TData, TValue>({
                       {header.isPlaceholder
                         ? null
                         : flexRender(
-                            header.column.columnDef.header,
-                            header.getContext()
-                          )}
+                          header.column.columnDef.header,
+                          header.getContext()
+                        )}
                     </TableHead>
                   )
                 })}
@@ -290,8 +83,8 @@ export function DataTable<TData, TValue>({
             ))}
           </TableHeader>
           <TableBody>
-            {table.getRowModel().rows?.length ? (
-              table.getRowModel().rows.map((row) => (
+            {dataTable.getRowModel().rows?.length ? (
+              dataTable.getRowModel().rows.map((row) => (
                 <TableRow
                   key={row.id}
                   data-state={row.getIsSelected() && "selected"}
@@ -320,12 +113,9 @@ export function DataTable<TData, TValue>({
         </Table>
       </div>
       <div className="space-y-2.5">
-        <DataTablePagination table={table} />
+        <DataTablePagination table={dataTable} />
         {floatingBar ? (
-          <DataTableFloatingBar
-            table={table}
-            deleteRowsAction={deleteRowsAction}
-          />
+          <DataTableFloatingBar table={dataTable}>{selectedActionControls}</DataTableFloatingBar>
         ) : null}
       </div>
     </div>
