+++ src/lib/catchError.ts
@@ -1,0 +1,15 @@
+import { toast } from "sonner";
+import { z } from "zod";
+
+export function catchError(err: unknown) {
+  if (err instanceof z.ZodError) {
+    const errors = err.issues.map((issue) => {
+      return issue.message;
+    });
+    return toast(errors.join("\n"));
+  } else if (err instanceof Error) {
+    return toast(err.message);
+  } else {
+    return toast("Something went wrong, please try again later.");
+  }
+}
